WIP: salt size reduction (104 -> 96) + epoch tag + counter elimination
Bulk uint104 -> uint96 across engine, manager, commit lib, and tests.
EIP-712 DualSignedReveal typehash updated to use uint96 revealerSalt.

New moveBuffer entry layout (256 bits tight pack):
  [p0Move 8 | p0Extra 16 | p0Salt 96 | p1Move 8 | p1Extra 16 | p1Salt 96 | epoch 16]

The 16-bit epoch is the low 16 bits of the battleKey OR'd with 1, so it's
non-zero and battle-unique (probabilistically). Used by executeBuffered
to detect "live for this battle" vs "stale leftover from a prior battle
that reused this storageKey but didn't drain its buffer." Replaces the
old bufferCounters SSTORE per submit -- removes ~5k gas of write traffic
per submitTurnMoves call (~70k per 14-turn game in production).

submitTurnMoves: writes only the entry slot (1 SSTORE), no counter.
executeBuffered: walks slots from engine.turnId until epoch mismatch.
bufferCounters mapping removed.

NOTE -- WIP commit before measurement. The 16-bit epoch has a 1/32768
collision risk between two battles ever using the same storageKey;
discussing whether to widen to 32 bits (cut salts to 88) before merge.

Build green, tests not yet run.

+++ src/IEngine.sol
@@ -41,18 +41,18 @@ interface IEngine {
         uint256 rng
     ) external returns (int32 damage, bytes32 eventType);
     function switchActiveMon(uint256 playerIndex, uint256 monToSwitchIndex) external;
-    function setMove(bytes32 battleKey, uint256 playerIndex, uint8 moveIndex, uint104 salt, uint16 extraData) external;
+    function setMove(bytes32 battleKey, uint256 playerIndex, uint8 moveIndex, uint96 salt, uint16 extraData) external;
     function execute(bytes32 battleKey) external returns (address winner);
     function executeWithMoves(
         bytes32 battleKey,
         uint8 p0MoveIndex,
-        uint104 p0Salt,
+        uint96 p0Salt,
         uint16 p0ExtraData,
         uint8 p1MoveIndex,
-        uint104 p1Salt,
+        uint96 p1Salt,
         uint16 p1ExtraData
     ) external returns (address winner);
-    function executeWithSingleMove(bytes32 battleKey, uint8 moveIndex, uint104 salt, uint16 extraData)
+    function executeWithSingleMove(bytes32 battleKey, uint8 moveIndex, uint96 salt, uint16 extraData)
         external
         returns (address winner);
     function executeBatchedTurns(bytes32 battleKey, uint256[] calldata entries)
