+++ test/StandardAttackPvPGasTest.sol
@@ -119,8 +119,8 @@ contract StandardAttackPvPGasTest is SignedCommitHelper {
         uint16 p1ExtraData
     ) internal {
         uint64 turnId = uint64(engine.getTurnIdForBattleState(battleKey));
-        uint104 committerSalt = uint104(uint256(keccak256(abi.encode("committer", battleKey, turnId))));
-        uint104 revealerSalt = uint104(uint256(keccak256(abi.encode("revealer", battleKey, turnId))));
+        uint96 committerSalt = uint96(uint256(keccak256(abi.encode("committer", battleKey, turnId))));
+        uint96 revealerSalt = uint96(uint256(keccak256(abi.encode("revealer", battleKey, turnId))));
 
         uint8 committerMoveIndex;
         uint16 committerExtraData;
