+++ test/mons/EkinekiTest.sol
@@ -473,7 +473,7 @@ contract EkinekiTest is Test, BattleHelper {
         // Alice forced switch to mon 2 (the one with savior complex)
         // After KO, playerSwitchForTurnFlag = 0 (Alice must switch, no commit needed)
         vm.startPrank(ALICE);
-        commitManager.revealMove(battleKey, SWITCH_MOVE_INDEX, uint104(0), uint16(2), true);
+        commitManager.revealMove(battleKey, SWITCH_MOVE_INDEX, uint96(0), uint16(2), true);
         engine.resetCallContext();
         // Verify that Alice's mon 2 got a sp atk boost (STAGE_1_BOOST = 15% of 100 = 15)
         int32 spAtkDelta = engine.getMonStateForBattle(battleKey, 0, 2, MonStateIndexName.SpecialAttack);
@@ -564,7 +564,7 @@ contract EkinekiTest is Test, BattleHelper {
 
         // Alice forced switch to mon 1 (savior complex triggers with 1 KO)
         vm.startPrank(ALICE);
-        commitManager.revealMove(battleKey, SWITCH_MOVE_INDEX, uint104(0), uint16(1), true);
+        commitManager.revealMove(battleKey, SWITCH_MOVE_INDEX, uint96(0), uint16(1), true);
         engine.resetCallContext();
         int32 spAtkDeltaFirstSwitch = engine.getMonStateForBattle(battleKey, 0, 1, MonStateIndexName.SpecialAttack);
         assertEq(spAtkDeltaFirstSwitch, 15, "Should get 15 sp atk boost from 1 KO");
@@ -583,7 +583,7 @@ contract EkinekiTest is Test, BattleHelper {
 
         // Alice forced switch back to mon 1 (savior complex should NOT trigger again)
         vm.startPrank(ALICE);
-        commitManager.revealMove(battleKey, SWITCH_MOVE_INDEX, uint104(0), uint16(1), true);
+        commitManager.revealMove(battleKey, SWITCH_MOVE_INDEX, uint96(0), uint16(1), true);
         engine.resetCallContext();
         int32 spAtkDeltaSecondSwitch = engine.getMonStateForBattle(battleKey, 0, 1, MonStateIndexName.SpecialAttack);
         // Boost is temp so it was cleared when mon 1 switched out, and savior complex
@@ -682,7 +682,7 @@ contract EkinekiTest is Test, BattleHelper {
 
         // Alice forced switch to mon 1
         vm.startPrank(ALICE);
-        commitManager.revealMove(battleKey, SWITCH_MOVE_INDEX, uint104(0), uint16(1), true);
+        commitManager.revealMove(battleKey, SWITCH_MOVE_INDEX, uint96(0), uint16(1), true);
         engine.resetCallContext();
         // Mon 1 has no ability, so no savior complex trigger
         // But the savior complex on mon 0 should NOT have been consumed (it didn't trigger)
